Refactoring code for K8s invoke command for later
